Refactor SqlConnectionDetails
