+++ src/OpenTelemetry.Instrumentation.SqlClient/Implementation/SqlActivitySourceHelper.cs
@@ -26,4 +26,35 @@ internal sealed class SqlActivitySourceHelper
     {
         new KeyValuePair<string, object?>(SemanticConventions.AttributeDbSystem, MicrosoftSqlServerDatabaseSystemName),
     };
+
+    public static void AddConnectionLevelDetailsToActivity(string dataSource, Activity activity, SqlClientTraceInstrumentationOptions options)
+    {
+        // TODO: The attributes added here are required. We need to consider
+        // collecting these attributes by default.
+        if (options.EnableConnectionLevelAttributes)
+        {
+            var connectionDetails = SqlConnectionDetails.ParseDataSource((string)dataSource);
+
+            // TODO: In the new conventions, instance name should now be captured
+            // as a part of db.namespace, when available.
+            if (options.EmitOldAttributes && !string.IsNullOrEmpty(connectionDetails.InstanceName))
+            {
+                activity.SetTag(SemanticConventions.AttributeDbMsSqlInstanceName, connectionDetails.InstanceName);
+            }
+
+            if (!string.IsNullOrEmpty(connectionDetails.ServerHostName))
+            {
+                activity.SetTag(SemanticConventions.AttributeServerAddress, connectionDetails.ServerHostName);
+            }
+            else
+            {
+                activity.SetTag(SemanticConventions.AttributeServerAddress, connectionDetails.ServerIpAddress);
+            }
+
+            if (connectionDetails.Port.HasValue)
+            {
+                activity.SetTag(SemanticConventions.AttributeServerPort, connectionDetails.Port);
+            }
+        }
+    }
 }
