+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlClientTests.cs
@@ -171,6 +171,59 @@ public class SqlClientTests : IDisposable
     }
 
     [Theory]
+    [InlineData(true, "localhost", "localhost", null, null, null)]
+    [InlineData(true, "127.0.0.1,1433", null, "127.0.0.1", null, null)]
+    [InlineData(true, "127.0.0.1,1434", null, "127.0.0.1", null, 1434)]
+    [InlineData(true, "127.0.0.1\\instanceName, 1818", null, "127.0.0.1", "instanceName", 1818)]
+    [InlineData(false, "localhost", null, null, null, null)]
+
+    // Test cases when EmitOldAttributes = false and EmitNewAttributes = true (i.e., OTEL_SEMCONV_STABILITY_OPT_IN=database)
+    [InlineData(true, "localhost", "localhost", null, null, null, false, true)]
+    [InlineData(true, "127.0.0.1,1433", null, "127.0.0.1", null, null, false, true)]
+    [InlineData(true, "127.0.0.1,1434", null, "127.0.0.1", null, 1434, false, true)]
+    [InlineData(true, "127.0.0.1\\instanceName, 1818", null, "127.0.0.1", null, 1818, false, true)]
+    [InlineData(false, "localhost", null, null, null, null, false, true)]
+
+    // Test cases when EmitOldAttributes = true and EmitNewAttributes = true (i.e., OTEL_SEMCONV_STABILITY_OPT_IN=database/dup)
+    [InlineData(true, "localhost", "localhost", null, null, null, true, true)]
+    [InlineData(true, "127.0.0.1,1433", null, "127.0.0.1", null, null, true, true)]
+    [InlineData(true, "127.0.0.1,1434", null, "127.0.0.1", null, 1434, true, true)]
+    [InlineData(true, "127.0.0.1\\instanceName, 1818", null, "127.0.0.1", "instanceName", 1818, true, true)]
+    [InlineData(false, "localhost", null, null, null, null, true, true)]
+    public void SqlClientAddsConnectionLevelAttributes(
+        bool enableConnectionLevelAttributes,
+        string dataSource,
+        string? expectedServerHostName,
+        string? expectedServerIpAddress,
+        string? expectedInstanceName,
+        int? expectedPort,
+        bool emitOldAttributes = true,
+        bool emitNewAttributes = false)
+    {
+        var activity = new Activity("Test Sql Activity");
+        var options = new SqlClientTraceInstrumentationOptions()
+        {
+            EnableConnectionLevelAttributes = enableConnectionLevelAttributes,
+            EmitOldAttributes = emitOldAttributes,
+            EmitNewAttributes = emitNewAttributes,
+        };
+        SqlActivitySourceHelper.AddConnectionLevelDetailsToActivity(dataSource, activity, options);
+
+        Assert.Equal(expectedServerHostName ?? expectedServerIpAddress, activity.GetTagValue(SemanticConventions.AttributeServerAddress));
+
+        if (emitOldAttributes)
+        {
+            Assert.Equal(expectedInstanceName, activity.GetTagValue(SemanticConventions.AttributeDbMsSqlInstanceName));
+        }
+        else
+        {
+            Assert.Null(activity.GetTagValue(SemanticConventions.AttributeDbMsSqlInstanceName));
+        }
+
+        Assert.Equal(expectedPort, activity.GetTagValue(SemanticConventions.AttributeServerPort));
+    }
+
+    [Theory]
     [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataWriteCommandError)]
     [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataWriteCommandError, false)]
     [InlineData(SqlClientDiagnosticListener.SqlDataBeforeExecuteCommand, SqlClientDiagnosticListener.SqlDataWriteCommandError, false, true)]
