+++ test/OpenTelemetry.Instrumentation.SqlClient.Tests/SqlConnectionDetailsTests.cs
@@ -1,0 +1,41 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using OpenTelemetry.Instrumentation.SqlClient.Implementation;
+using Xunit;
+
+namespace OpenTelemetry.Instrumentation.SqlClient.Tests;
+
+public class SqlConnectionDetailsTests
+{
+    [Theory]
+    [InlineData("localhost", "localhost", null, null, null)]
+    [InlineData("127.0.0.1", null, "127.0.0.1", null, null)]
+    [InlineData("127.0.0.1,1433", null, "127.0.0.1", null, null)]
+    [InlineData("127.0.0.1, 1818", null, "127.0.0.1", null, 1818)]
+    [InlineData("127.0.0.1  \\  instanceName", null, "127.0.0.1", "instanceName", null)]
+    [InlineData("127.0.0.1\\instanceName, 1818", null, "127.0.0.1", "instanceName", 1818)]
+    [InlineData("tcp:127.0.0.1\\instanceName, 1818", null, "127.0.0.1", "instanceName", 1818)]
+    [InlineData("tcp:localhost", "localhost", null, null, null)]
+    [InlineData("tcp : localhost", "localhost", null, null, null)]
+    [InlineData("np : localhost", "localhost", null, null, null)]
+    [InlineData("lpc:localhost", "localhost", null, null, null)]
+    [InlineData("np:\\\\localhost\\pipe\\sql\\query", "localhost", null, null, null)]
+    [InlineData("np : \\\\localhost\\pipe\\sql\\query", "localhost", null, null, null)]
+    [InlineData("np:\\\\localhost\\pipe\\MSSQL$instanceName\\sql\\query", "localhost", null, "instanceName", null)]
+    public void ParseDataSourceTests(
+        string dataSource,
+        string? expectedServerHostName,
+        string? expectedServerIpAddress,
+        string? expectedInstanceName,
+        int? expectedPort)
+    {
+        var sqlConnectionDetails = SqlConnectionDetails.ParseDataSource(dataSource);
+
+        Assert.NotNull(sqlConnectionDetails);
+        Assert.Equal(expectedServerHostName, sqlConnectionDetails.ServerHostName);
+        Assert.Equal(expectedServerIpAddress, sqlConnectionDetails.ServerIpAddress);
+        Assert.Equal(expectedInstanceName, sqlConnectionDetails.InstanceName);
+        Assert.Equal(expectedPort, sqlConnectionDetails.Port);
+    }
+}
